reorganized code better
