+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/BlueJewelOnly.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Autons;
+package org.firstinspires.ftc.teamcode.Autonomous;
 
 /**
  * Created by anikaitsingh on 12/9/17.
