+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/JewelAutonRed.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Autons;
+package org.firstinspires.ftc.teamcode.Autonomous;
 
 
 import android.graphics.Color;
